Added CLI options and renamed some files
